[API] route 삭제 오류 수정

+++ app/RouteLike.php
@@ -71,7 +71,9 @@ class RouteLike extends Model
             ->get();
 
         $pickId = $idCheck->pluck('id');
-
+        if ($pickId->isEmpty()) {
+            return [];
+        }
         return self::where('id', $pickId)->delete();
     }
 }
